Mini Project in C - Scientific Calculator
About the Project

The purpose of this project was to develop a program in such a way where by it can analyze and make calculations using arithmetical algorithms that are possible to be done on most scientific calculator. Due to the programming language used, we will also be able to comment and verify our functions written in the program with REAL values that would automatically appear on a calculator display.
The project has three main parts:

1. The first part of the project is to develop and implement a very basic mathematical function such as addition, subtraction, multiplication, division and integer splicing in an easy to understand way using only a few simple lines of instructions. The best way to do this would be by using the Visual Basic programming language as we would be able to see our functions as we program them. This should also let us verify that our functions are working properly when we run them on a real calculator. Once these bare minimum mathematical functions have been developed, we will next expand on this code and make more complex calculations such as square root.

2. The second part of the project is to develop the more complex mathematical functions such as square root, exponential, logarithms etc., using our existing arithmetic code developed in Part 1. This will require us to make use of the mathematical concepts taught to us in class, including:

a) Trigonometry : involves calculating trigonometric functions. 
b) Exponential and Logarithm: both of these involve raising a number to a power or base (e.g., 2^5) 
c) Integration and derivatives : integration is finding an antiderivative while derivatives is a function's rate of change with respect to another function (such as movement on an x-y axis).

3. The last part of the project will be to create programs that use these mathematical functions, which will require us to think of different situations in which we would want to use these functions in real life. As well as thinking about how we would want our programs to work and what parameters we should make available for the user for inputting data into our calculator program. For example, being able to enter a value from a scale on a scientific calculator into our program and make it compute things like trigonometric functions, square root etc., as opposed to entering values manually using buttons on the calculator itself.